requirements.txt
requests
beautifulsoup4
pdfplumber
pandas
psycopg[binary]
python-dotenv
typer